+++ changelog.md
@@ -8,6 +8,32 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+<!--
+## vNEXT - DATE
+
+### Notes
+
+### Added
+
+### Changed
+
+### Removed
+
+### Other work
+-->
+
+## v2.0.1 - WIP
+
+### Notes
+
+### Added
+
+### Changed
+
+### Removed
+
+### Other work
+
 ## v2.0.0 - 2026-07-13
 
 ### Added